ed448-goldilocks: don't rename `signature` crate
Uses the original crate name and re-exports it under
`ed448_goldilocks::sign::signature`
