+++ ed448-goldilocks/Cargo.toml
@@ -16,21 +16,23 @@ This crate also includes signing and verifying of Ed448 signatures.
 """
 
 [dependencies]
-crypto_signature = { version = "3.0.0-rc.0", default-features = false, features = ["digest", "rand_core"], optional = true, package = "signature" }
 elliptic-curve = { version = "0.14.0-rc.6", features = ["arithmetic", "hash2curve", "pkcs8"] }
-ed448 = { version = "=0.5.0-pre.0", default-features = false, optional = true }
 rand_core = { version = "0.9", default-features = false }
-serdect = { version = "0.3.0", optional = true }
 sha3 = { version = "0.11.0-rc.0", default-features = false }
 subtle = { version = "2.6", default-features = false }
 
+# optional dependencies
+ed448 = { version = "=0.5.0-pre.0", optional = true, default-features = false }
+serdect = { version = "0.3.0", optional = true }
+signature = { version = "3.0.0-rc.0", optional = true, default-features = false, features = ["digest", "rand_core"] }
+
 [features]
 default = ["std", "signing", "pkcs8"]
-alloc = ["crypto_signature/alloc", "ed448?/alloc", "elliptic-curve/alloc", "serdect/alloc"]
+alloc = ["ed448?/alloc", "elliptic-curve/alloc", "serdect?/alloc", "signature?/alloc"]
 std = ["alloc"]
 bits = ["elliptic-curve/bits"]
 pkcs8 = ["ed448?/pkcs8", "elliptic-curve/pkcs8"]
-signing = ["dep:crypto_signature", "ed448"]
+signing = ["dep:ed448", "dep:signature"]
 serde = ["dep:serdect", "ed448?/serde_bytes"]
 
 [dev-dependencies]
